pyproject.toml
[project]
name = "agente-ia-zero"
version = "0.1.0"
description = "Projeto teste para um agente de IA simples."
authors = [
    {name = "ruizgabriel161",email = "ruizgabriel161@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain (>=1.0.5,<2.0.0)",
    "langchain-chroma (>=1.0.0,<2.0.0)",
    "chromadb (>=1.3.4,<2.0.0)",
    "pypdf (>=6.2.0,<7.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-text-splitters (>=1.0.0,<2.0.0)",
    "langchain-google-genai (>=3.0.2,<4.0.0)",
    "langchain-ollama (>=1.0.0,<2.0.0)",
    "langgraph (>=1.0.3,<2.0.0)",
    "langgraph-cli[inmem] (>=0.4.7,<0.5.0)"
]

[tool.poetry]
packages = [{include = "agente_ia_zero", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.0"
pytest-cov = "^7.0.0"
taskipy = "^1.14.1"
ruff = "^0.14.4"
pyright = "^1.1.407"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79

[tool.ruff.lint]
preview = true

[tool.ruff.format]
preview = true

[tool.pytest.ini_options]
pythonpath = "."


# ============================
# Tipagem (Pyright)
# ============================
[tool.pyright]
typeCheckingMode = "basic"
pythonVersion = "3.12"
include = ["src", 'tests']
exclude = [
    "**/venv",
    "**/.venv",
    "**/env",
    "**/.env",
    "**/node_modules",
    "**/__pycache__",
]
